lesson26_flask/flask_email_students/requirements.txt
attrs==22.1.0
boto3==1.26.26
botocore==1.29.26
cached-property==1.5.2
certifi==2022.9.24
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
datapackage==1.15.2
dominate==2.7.0
et-xmlfile==1.1.0
Flask==2.2.2
Flask-Bootstrap==3.3.7.1
Flask-Moment==1.0.5
flask-paginate==2022.1.8
Flask-WTF==1.0.1
greenlet==2.0.1
idna==3.4
ijson==3.1.4
importlib-metadata==5.1.0
importlib-resources==5.10.1
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonlines==3.1.0
jsonpointer==2.3
jsonschema==4.17.3
linear-tsv==1.1.0
MarkupSafe==2.1.1
openpyxl==3.0.10
packaging==21.3
peewee==3.15.4
pkgutil_resolve_name==1.3.10
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
python-dotenv==0.21.0
requests==2.28.1
rfc3986==2.0.0
s3transfer==0.6.0
six==1.16.0
SQLAlchemy==1.4.44
tableschema==1.20.2
tabulator==1.53.5
unicodecsv==0.14.1
urllib3==1.26.13
visitor==0.1.3
Werkzeug==2.2.2
WTForms==3.0.1
xlrd==2.0.1
zipp==3.11.0
